Add stub replacement for x86 decoder
This also immediately decodes operands. The only place doDelayedDecode
is used is in Instruction's DECODE_OPCODE. Moreover, everywhere in
Dyninst that an InstructionDecoder is used, the operands are immediately
needed. There's no need to pay the expense of invoking the decoder and
then immediately invoking it again with a different set of arguments.