Update instPoint::liveRegisters
The previous check against `getAllRegs().size()` is equivalent to
checking if `liveRegs_` has been written to which is only done here
and instPoint::fork. The latter is really more of a copy constructor,
so it's not doing any logic to populate it.

This single-initialization should be fine because the objects returned
by `func()` and `type()` are class invariants, so repeated calls to
`liveRegisters` would just recalculate the same query which is
wasteful.