Remove segment registers
PowerPC ISA v2.01
E.25 Segment Register Manipulation Instructions

These can never be referred to by name. Rather, they are implicitly used
by instructions like 'mfsrr2' which is itself an alias for
'mfspr r0, 0x3de'. Both binutils and Capstone decode it this way and
mark the read "register" as the immediate 0x3de. They are not part of
the ABI, so no pseudoregister is needed.