Remove Software-Use SPRs
3.3.3 Software-Use SPRs, ISA v2.01, Book III

sprg0-2 are always privileged. sprg3 can be used in both privileged
and user mode. sprg4-7 are available in user mode. However, these
can never be referred to by name. Rather, they are implicitly used
by instructions like 'mfsprg4' which is itself an alias for
'mfspr r0, 0x104'. Both binutils and Capstone decode it this way and
mark the read "register" as the immediate 0x104. They are not part of
the ABI, so no pseudoregister is needed.