Remove time base registers
7.2.1 Writing the Time Base, ISA v2.07, Book III

These can never be referred to by name. Rather, they are implicitly used
by instructions like 'mttbl' which is itself an alias for
'mtspr 0x11c, r0'. Both binutils and Capstone decode it this way and
mark the read "register" as the immediate 0x11c. They are not part of
the ABI, so no pseudoregister is needed.