Remove Data Storage Interrupt Status Register
6.2.5 Data Storage Interrupt Status Register, ISA v2.07, Book III

This is a privileged register and can never be referred to by name.
Rather, it is implicitly used by instructions like 'mfdsisr r4' which
is itself an alias for 'mfspr 0x12, r4'. Both binutils and Capstone
decode it this way and mark the read "register" as the immediate 0x12.
It is not part of the ABI, so no pseudoregister is needed.