Remove Data Address Register
6.2.3 Data Address Register, ISA v2.07, Book III

This can never be referred to by name and is a privileged register.
Rather, it is implicitly used by instructions like 'mfdar r4' which
is itself an alias for 'mfspr 0x13, r4'. Both binutils and Capstone
decode it this way and mark the read "register" as the immediate 0x13.
It is not part of the ABI, so no pseudoregister is needed.