Remove Storage Description Register 1
5.7.7.2 Storage Description Register 1, ISA v2.07, Book III

This can never be referred to by name and is only available in
hypervisor mode. Rather, it is implicitly used by instructions like
'mfsdr1 r4' which is itself an alias for 'mfspr 0x19, r4'. Both binutils
and Capstone decode it this way and mark the read "register" as the
immediate 0x19. It is not part of the ABI, so no pseudoregister is
needed.