Remove Machine Status Save/Restore Registers
6.2.1 Machine Status Save/Restore Registers, ISA v2.07, Book III

These can never be referred to by name and are privileged registers.
Rather, they are implicitly used by instructions like 'mfsrr0 r4' which
is itself an alias for 'mfspr 0x1a, r4'. Both binutils and Capstone
decode it this way and mark the read "register" as the immediate 0x1a.
They are not part of the ABI, so no pseudoregister is needed.