Remove Processor Version Register
4.3.1 Processor Version Register, ISA v2.07, Book III

This can never be referred to by name and is a privileged register.
Rather, it is implicitly used by instructions like 'mfpvr r4' which
is itself an alias for 'mfspr 0x11f, r4'. Both binutils and Capstone
decode it this way and mark the read "register" as the immediate 0x11f.
They are not part of the ABI, so no pseudoregister is needed.