Remove Interrupt Vector Prefix Register
7.2.11 Interrupt Vector Prefix Register, ISA v2.07, Book III

This can never be referred to by name and is a privileged register.
Rather, it is implicitly used by instructions like 'mfivpr r4' which
is itself an alias for 'mfspr 0x3f, r4'. Both binutils and Capstone
decode it this way and mark the read "register" as the immediate 0x3f.
They are not part of the ABI, so no pseudoregister is needed.