Remove Program Priority Registers
3.1 Program Priority Registers, ISA v2.07, Book II

These can never be referred to by name but are not privileged registers.
Rather, they are implicitly used by instructions like 'mfppr r4' which
is itself an alias for 'mfspr 0x380, r4'. Both binutils and Capstone
decode it this way and mark the read "register" as the immediate 0x380.
They are not part of the ABI, so no pseudoregister is needed.