Fix register calculation in parse_func::calcUsedRegs for aarch64
Ultimately, these are consumed by EmitterAARCH64::clobberAllFuncCall
to populate values in a registerSlot that uses aarch64Registers_t to
represent registers. That enum and the corresponding conversion map in
RegisterConversion-aarch64.C use the base registers x<N> for GPRs and
q<N> for FPRs, so the registers are correspondingly transformed here
before being stored in parse_func::usedRegisters.