Replace dyninst_lock_t with mtx_t from C11
The x86 lock was fine, but the aarch64 spinlock suffers from the
standard spinlock problems. I'm not sure about the ppc ones. In any
case, this makes a single implementation for all platforms and removes
the maintenance burden when porting to new architectures.

The indirection in dyninst_lock_t::mutex isn't ideal but is needed
because there is a global namespace conflict with 'call_once' between
stdatomic.h, <atomic>, and Boost.Threads when dyninstRTExport.h is
included in a C++ source file.